Add strict mypy type checking across the whole library
Expand the mypy config from the diag submodule to all of src/QuantumSCC
(strict mode) and add sympy/IPython to the ignore-missing-imports override
alongside scipy.

Annotate every module (circuit, core.quantization, core.topology,
core.geometry, core.elements, utils.linalg, utils.symbolic, model): return
types, parameters, and var-annotated locals. Use TYPE_CHECKING imports in
model.py and symbolic.py to avoid import cycles.

Fix genuine type issues surfaced by the checker (all behavior-preserving):
- elements.py: wrap np.random.normal(...)[0] returns in float()
- topology.py: type self.elements as list[list[Any]], resolving invalid
  ndarray index types
- quantization.py: rename the `for k, v in _sup.items()` loop vars to
  pat/rep (k was previously bound as int, a real str/int clash); mark
  IPython's injected get_ipython() with type: ignore[name-defined]

The degenerate-eigenvalue F821 branch in linalg.py stays intentionally
broken: documented type: ignore[has-type] lets mypy pass while the FIXME
markers remain and ruff still reports the underlying undefined-name bug.

mypy: clean (15 files). pytest: 388 passed.

+++ src/QuantumSCC/core/elements.py
@@ -70,7 +70,7 @@ class Capacitor:
             return cMean
         else:
             # Simple fabrication error model
-            return np.random.normal(cMean, cMean * getattr(self, 'error', 0) / 100, 1)[0]
+            return float(np.random.normal(cMean, cMean * getattr(self, 'error', 0) / 100, 1)[0])
 
     def energy(self) -> float:
         """
@@ -144,7 +144,7 @@ class Inductor:
         if not random:
             return lMean
         else:
-            return np.random.normal(lMean, lMean * getattr(self, 'error', 0) / 100, 1)[0]
+            return float(np.random.normal(lMean, lMean * getattr(self, 'error', 0) / 100, 1)[0])
 
     def energy(self) -> float:
         """
